feat: implement "break"
The condition that it is only used inside a loop
is being tracked in Parser vis loopDepth variable.

The implementation is done in Interpreter by setting 'breaking' flag. This makes it so all remaining statements in
a block are skipped. The flag is reset at the end of a While loop.

Chapter: 9 challenge 3